Subo cambios de algunos fixes

+++ common/includes/protocol.h
@@ -85,7 +85,7 @@ struct MsgInteract {
 struct MsgTrade {
     uint8_t opcode;  // BUY_ITEM or SELL_ITEM
     uint32_t npc_id;
-    uint16_t item_id;
+    uint8_t item_id;
     uint16_t quantity;
 };
